move inline code from renf_elem.h to proper implementation files
